requirements.txt
pandas
numpy
tqdm
python-dateutil
streamlit
fastapi
uvicorn
google-generativeai
python-dotenv
